Separate CredentialsForm to separate component
