+++ Source/SelfService/Web/integrations/connection/connectionDetails/exposeData/Credentials/CredentialsSection.tsx
@@ -9,6 +9,7 @@ import { AlertBox, Button, ContentSection } from '@dolittle/design-system';
 import { useConnectionsIdServiceAccountsGet } from '../../../../../apis/integrations/serviceAccountApi.hooks';
 import { useConnectionId } from '../../../../../integrations/routes.hooks';
 import { CredentialsList } from './CredentialsList';
+import { GenerateCredentialsForm } from './GenerateCredentialsForm';
 
 export type CredentialsSectionProps = {};
 
@@ -25,12 +26,7 @@ export const CredentialsSection = (props: CredentialsSectionProps) => {
     if (isError) {
         return <AlertBox message={`Error while fetching credentials list. ${error}`} />;
     }
-    const credentialToken = 'n$H8rAp3mDJGiR7Adn4@paAzQ7J$cNJSEkzqPDYi';
 
-    const handleTokenCopy = () => {
-        navigator.clipboard.writeText(credentialToken);
-        enqueueSnackbar('Token copied to clipboard.');
-    };
 
 
     return (
@@ -39,38 +35,9 @@ export const CredentialsSection = (props: CredentialsSectionProps) => {
             headerProps={{ buttons: [{ label: 'Generate New Credentials', variant: 'outlined', onClick: () => setOpenCredentials(true) }] }}
         >
             <Collapse in={openCredentials}>
-                <Grid container spacing={3} sx={{ pb: 5 }}>
-                    <Grid item>
-                        <Typography sx={{ mb: 2 }}>Who or what are these credentials for?</Typography>
-                        <TextField id='credentialsName' label='Name' size='small' />
-                    </Grid>
-
-                    <Grid item>
-                        <Typography sx={{ mb: 2 }}>Credential Token</Typography>
-                        <TextField
-                            id='credentialsToken'
-                            label='Token'
-                            size='small'
-                            type='password'
-                            value={credentialToken}
-                            disabled
-                            sx={{ width: 400 }}
-                        />
-                        <FormHelperText>This bearer token should be used in the request header.</FormHelperText>
-                    </Grid>
-
-                    <Grid container item spacing={3} xs={6} sx={{ alignItems: 'center', mt: 0 }}>
-                        <Grid item>
-                            <Button label='Copy Token' startWithIcon='CopyAllRounded' onClick={handleTokenCopy} />
-                        </Grid>
-
-                        <Grid item>
-                            <Button label='Delete credentials' startWithIcon='DeleteRounded' onClick={() => setOpenCredentials(false)} />
-                        </Grid>
-                    </Grid>
-                </Grid>
-                <CredentialsList credentials={data || []} />
+                <GenerateCredentialsForm onFormComplete={() => setOpenCredentials(false)} />
             </Collapse>
+                <CredentialsList credentials={data || []} />
         </ContentSection>
     );
 };
